requirements.txt
livekit-agents[deepgram,openai,silero,turn-detector]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv
psycopg2-binary